requirements.txt
numpy
pandas
scikit-learn
yfinance
alpaca-py
PyPortfolioOpt
pyarrow
pytz
pandas_market_calendars
boto3
python-dotenv
